srcs/backend/soltyback/pyproject.toml
[project]
name = "soltyback"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=44.0.2",
    "django-rest-framework>=0.1.0",
    "django>=5.2",
    "pillow>=11.1.0",
    "solana>=0.36.6",
    "black>=25.1.0",
    "requests>=2.32.3",
]
